Ingest instagram to the real-time table format.
Changes to ingest July data and to generate data for entity alignment.
